refactor: move template vars into a separate file
