+++ web-prover-core/src/http.rs
@@ -19,7 +19,7 @@ use serde::{Deserialize, Serialize};
 use tracing::debug;
 pub use web_proof_circuits_witness_generator::json::JsonKey;
 
-use crate::errors::ManifestError;
+use crate::{errors::ManifestError, template, template::TemplateVar};
 
 /// Max HTTP headers
 pub const MAX_HTTP_HEADERS: usize = 25;
@@ -302,12 +302,9 @@ impl NotaryResponse {
 }
 
 /// Default HTTP version
-pub fn default_version() -> String { HTTP_1_1.to_string() }
+fn default_version() -> String { HTTP_1_1.to_string() }
 /// Default HTTP message
-pub fn default_message() -> String { "OK".to_string() }
-
-/// Returns an empty `HashMap` as the default value for `vars`
-fn default_empty_vars() -> HashMap<String, TemplateVar> { HashMap::new() }
+fn default_message() -> String { "OK".to_string() }
 
 /// HTTP Response items required for circuits
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -339,21 +336,19 @@ impl ManifestResponse {
   ///
   /// The validated HTTP response.
   pub fn validate(&self) -> Result<(), ManifestError> {
-    // TODO: What are legal statuses?
+    // We only support 200 and 201
     const VALID_STATUSES: [&str; 2] = ["200", "201"];
     if !VALID_STATUSES.contains(&self.status.as_str()) {
       return Err(ManifestError::InvalidManifest("Unsupported HTTP status".to_string()));
     }
-
-    // TODO: What HTTP versions are supported?
+    // We only support HTTP/1.1
     if self.version != "HTTP/1.1" {
       return Err(ManifestError::InvalidManifest(
         "Invalid HTTP version: ".to_string() + &self.version,
       ));
     }
 
-    // TODO: What is the max supported message length?
-    // TODO: Not covered by serde's #default annotation. Is '""' a valid message?
+    // An empty message is not allowed
     if self.message.len() > 1024 || self.message.is_empty() {
       return Err(ManifestError::InvalidManifest(
         "Invalid message length: ".to_string() + &self.message,
@@ -400,22 +395,8 @@ impl ManifestResponse {
   }
 }
 
-/// Template variable type
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct TemplateVar {
-  /// Optional description for the end user
-  pub description: Option<String>,
-  /// Indicates if the value is required
-  #[serde(default = "default_required")]
-  pub required:    bool,
-  /// A default value, must be set when `required` is false
-  pub default:     Option<String>,
-  /// Regex pattern for validation of user input
-  pub pattern:     Option<String>,
-}
-
-/// Default value for the required field is true
-fn default_required() -> bool { true }
+/// Returns an empty `HashMap` as the default value for `vars`
+fn default_empty_vars() -> HashMap<String, TemplateVar> { HashMap::new() }
 
 /// HTTP Request items required for circuits
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -472,78 +453,42 @@ impl ManifestRequest {
     Ok(())
   }
 
-  pub fn validate_vars(&self) -> Result<(), ManifestError> {
-    let mut all_tokens = vec![];
+  fn collect_tokens(&self) -> Vec<String> {
+    let mut all_tokens = Vec::new();
 
-    // Parse and validate tokens in the body
-    if let Some(body_tokens) = self.body.as_ref().map(extract_tokens) {
-      for token in &body_tokens {
-        if !self.vars.contains_key(token) {
-          return Err(ManifestError::InvalidManifest(format!(
-            "Token `<% {} %>` not declared in `vars`",
-            token
-          )));
-        }
-      }
+    // Collect tokens from body
+    if let Some(body_tokens) = self.body.as_ref().map(template::extract_tokens) {
       all_tokens.extend(body_tokens);
     }
 
-    // Parse and validate tokens in headers
+    // Collect tokens from headers
     for value in self.headers.values() {
-      let header_tokens = extract_tokens(&serde_json::Value::String(value.clone()));
-      for token in &header_tokens {
-        if !self.vars.contains_key(token) {
-          return Err(ManifestError::InvalidManifest(format!(
-            "Token `<% {} %>` not declared in `vars`",
-            token
-          )));
-        }
-      }
+      let header_tokens = template::extract_tokens(&serde_json::Value::String(value.clone()));
       all_tokens.extend(header_tokens);
     }
 
-    // Validate each `vars` entry
-    for (key, var_def) in &self.vars {
-      // Check if the variable is used in the template
+    all_tokens
+  }
+
+  fn validate_tokens(&self, tokens: &[String]) -> Result<(), ManifestError> {
+    for token in tokens {
+      if !self.vars.contains_key(token) {
+        return Err(ManifestError::InvalidManifest(format!(
+          "Token `<% {} %>` not declared in `vars`",
+          token
+        )));
+      }
+    }
+    Ok(())
+  }
+
+  pub fn validate_vars(&self) -> Result<(), ManifestError> {
+    let all_tokens = self.collect_tokens();
+    self.validate_tokens(&all_tokens)?;
+
+    for (key, variable) in &self.vars {
       let is_used = all_tokens.contains(key);
-
-      // If the variable is required but not used, return an error
-      if var_def.required && !is_used {
-        return Err(ManifestError::InvalidManifest(format!(
-          "Required variable `{}` is not used in the template",
-          key
-        )));
-      }
-
-      // If the variable is not required, it must have a default value
-      if !var_def.required && var_def.default.is_none() {
-        return Err(ManifestError::InvalidManifest(format!(
-          "Non-required variable `{}` must have a default value",
-          key
-        )));
-      }
-
-      // Validate pattern (if defined)
-      if let Some(pattern) = var_def.pattern.as_ref() {
-        // Using `regress` crate for compatibility with ECMAScript regular expressions
-        let _regex = regress::Regex::new(pattern).map_err(|_| {
-          ManifestError::InvalidManifest(format!("Invalid regex pattern for `{}`", key))
-        })?;
-
-        // If there's a default value, validate it against the pattern
-        if let Some(default_value) = var_def.default.as_ref() {
-          let regex = regex::Regex::new(pattern).map_err(|_| {
-            ManifestError::InvalidManifest(format!("Invalid regex pattern for `{}`", key))
-          })?;
-
-          if !regex.is_match(default_value) {
-            return Err(ManifestError::InvalidManifest(format!(
-              "Default value for `{}` does not match the specified pattern",
-              key
-            )));
-          }
-        }
-      }
+      variable.validate(key, is_used)?;
     }
 
     Ok(())
@@ -644,37 +589,6 @@ impl ManifestRequest {
     // self.method == other.method && self.url == other.url && self.body == other.body
     self.method == other.method && self.url == other.url
   }
-}
-
-fn extract_tokens(value: &serde_json::Value) -> Vec<String> {
-  let mut tokens = vec![];
-
-  match value {
-    serde_json::Value::String(s) => {
-      let token_regex = regex::Regex::new(r"<%\s*(\w+)\s*%>").unwrap();
-      for capture in token_regex.captures_iter(s) {
-        if let Some(token) = capture.get(1) {
-          // Extract token name
-          tokens.push(token.as_str().to_string());
-        }
-      }
-    },
-    serde_json::Value::Object(map) => {
-      // Recursively parse nested objects
-      for (_, v) in map {
-        tokens.extend(extract_tokens(v));
-      }
-    },
-    serde_json::Value::Array(arr) => {
-      // Recursively parse arrays
-      for v in arr {
-        tokens.extend(extract_tokens(v));
-      }
-    },
-    _ => {},
-  }
-
-  tokens
 }
 
 #[cfg(test)]
