+++ web-prover-core/src/lib.rs
@@ -3,4 +3,5 @@ pub mod http;
 pub mod manifest;
 
 pub mod proof;
+mod template;
 pub mod test_utils;
